add data check to imsave

+++ cvutils/imsave.py
@@ -11,7 +11,7 @@ def imsave(image: np.ndarray, path: str) -> None:
         assert image_max <= 1.0 or image_min >= 0.0, \
             f'image value (float) of out of range, image.max()={image_max},'\
             f' image.min()={image_min}.'
-        image = (image * 255).type(np.int)
+        image = (image * 255).astype(np.uint8)
     else:
         # dtype of image is int
         image_max = image.max()
@@ -19,6 +19,7 @@ def imsave(image: np.ndarray, path: str) -> None:
         assert image_max <= 255 or image_min >= 0, \
             f'image value (int) of out of range, image.max()={image_max},'\
             f' image.min()={image_min}.'
+        image = image.astype(np.uint8)
 
     image = cv2.cvtColor(image, cv2.COLOR_RGB2BGR)
     cv2.imwrite(path, image)
